+++ src/cpp/dart_helpers.h
@@ -13,12 +13,19 @@
 #define DART_CHECK_RET(var, expr, ret, message)                                                                        \
   Dart_Handle var = (expr);                                                                                            \
   if (Dart_IsError(var)) {                                                                                             \
-    GD_PRINT_ERROR("GodotDart: " message ": ");                                                                      \
+    GD_PRINT_ERROR("GodotDart: " message ": ");                                                                        \
     GD_PRINT_ERROR(Dart_GetError(var));                                                                                \
     return ret;                                                                                                        \
   }
 
 #define DART_CHECK(var, expr, message) DART_CHECK_RET(var, expr, , message)
+
+#define DART_HANDLE_ERROR(var, message)                                                                                \
+  if (Dart_IsError(var)) {                                                                                             \
+    GD_PRINT_ERROR("GodotDart: " message ": ");                                                                        \
+    GD_PRINT_ERROR(Dart_GetError(var));                                                                                \
+    return;                                                                                                            \
+  }
 
 class DartBlockScope {