+++ src/dart/godot_dart_build/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.0
+
+- Add support for Godot Global Classes
+
 ## 0.1.0
 
 - Initial version.